requirements_version.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.68.1
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
dydantic==0.0.8
emoji==2.15.0
executing==2.2.1
fastapi==0.118.0
filetype==1.2.0
frozenlist==1.7.0
greenlet==3.2.4
groq==0.32.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.20
langchain-community==0.3.30
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langmem==0.0.29
langsmith==0.4.31
lxml==6.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.4
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.3
olefile==0.47
openai==1.109.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traitlets==5.14.3
trustcall==0.0.39
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
